requirements.txt
kafka-python==2.0.2
google-cloud-bigquery==3.11.4
python-dotenv==1.0.0